+++ libs/providers/config-cat/src/lib/config-cat-provider.ts
@@ -1,82 +1,78 @@
 import {
   EvaluationContext,
-  GeneralError,
   JsonValue,
   OpenFeatureEventEmitter,
-  ParseError,
   Provider,
   ProviderEvents,
-  ProviderStatus,
   ResolutionDetails,
-  ResolutionReason,
-  StandardResolutionReasons,
+  Paradigm,
+  ProviderNotReadyError,
   TypeMismatchError,
+  FlagNotFoundError,
+  ParseError,
 } from '@openfeature/server-sdk';
-import { getClient, IConfig, IConfigCatClient, IEvaluationDetails, SettingValue } from 'configcat-js-ssr';
-import { transformContext } from './context-transformer';
+import {
+  isType,
+  PrimitiveType,
+  PrimitiveTypeName,
+  toResolutionDetails,
+  transformContext,
+} from '@openfeature/config-cat-core';
+import { PollingMode } from 'configcat-common';
+import { IConfigCatClient, getClient, IConfig, OptionsForPollingMode } from 'configcat-node';
 
 export class ConfigCatProvider implements Provider {
   public readonly events = new OpenFeatureEventEmitter();
-  private _status: ProviderStatus = ProviderStatus.NOT_READY;
-  private readonly _clientParameters: Parameters<typeof getClient>;
+  private readonly _clientFactory: (provider: ConfigCatProvider) => IConfigCatClient;
+  private _hasError = false;
   private _client?: IConfigCatClient;
+
+  public runsOn: Paradigm = 'server';
 
   public metadata = {
     name: ConfigCatProvider.name,
   };
 
-  constructor(...params: Parameters<typeof getClient>) {
-    this._clientParameters = params;
+  protected constructor(clientFactory: (provider: ConfigCatProvider) => IConfigCatClient) {
+    this._clientFactory = clientFactory;
   }
 
-  public static create(...params: Parameters<typeof getClient>) {
-    return new ConfigCatProvider(...params);
-  }
-
-  get status() {
-    return this._status;
-  }
-
-  public async initialize(): Promise<void> {
-    return new Promise((resolve) => {
-      const originalParameters = this._clientParameters;
-      originalParameters[2] ??= {};
-
-      const options = originalParameters[2];
-      const oldSetupHooks = options.setupHooks;
+  public static create<TMode extends PollingMode>(
+    sdkKey: string,
+    pollingMode?: TMode,
+    options?: OptionsForPollingMode<TMode>,
+  ): ConfigCatProvider {
+    // Let's create a shallow copy to not mess up caller's options object.
+    options = options ? { ...options } : ({} as OptionsForPollingMode<TMode>);
+    return new ConfigCatProvider((provider) => {
+      const oldSetupHooks = options?.setupHooks;
 
       options.setupHooks = (hooks) => {
         oldSetupHooks?.(hooks);
 
-        // After resolving, once, we can simply emit events the next time
-        const onProviderReady = () => {
-          this.events.emit(ProviderEvents.Ready);
-          this._status = ProviderStatus.READY;
-        };
-
-        hooks.once('clientReady', () => {
-          hooks.on('clientReady', onProviderReady);
-          onProviderReady();
-          resolve();
-        });
-
         hooks.on('configChanged', (projectConfig: IConfig | undefined) =>
-          this.events.emit(ProviderEvents.ConfigurationChanged, {
+          provider.events.emit(ProviderEvents.ConfigurationChanged, {
             flagsChanged: projectConfig ? Object.keys(projectConfig.settings) : undefined,
           }),
         );
 
         hooks.on('clientError', (message: string, error) => {
-          this._status = ProviderStatus.ERROR;
-          this.events.emit(ProviderEvents.Error, {
+          provider._hasError = true;
+          provider.events.emit(ProviderEvents.Error, {
             message: message,
             metadata: error,
           });
         });
       };
 
-      this._client = getClient(...originalParameters);
+      return getClient(sdkKey, pollingMode, options);
     });
+  }
+
+  public async initialize(): Promise<void> {
+    const client = this._clientFactory(this);
+    await client.waitForReady();
+    this._client = client;
   }
 
   public get configCatClient() {
@@ -92,21 +88,7 @@ export class ConfigCatProvider implements Provider {
     defaultValue: boolean,
     context: EvaluationContext,
   ): Promise<ResolutionDetails<boolean>> {
-    if (!this._client) {
-      throw new GeneralError('Provider is not initialized');
-    }
-
-    const { value, ...evaluationData } = await this._client.getValueDetailsAsync<SettingValue>(
-      flagKey,
-      undefined,
-      context.targetingKey != null ? transformContext(context) : undefined,
-    );
-
-    const validatedValue = validateFlagType('boolean', value);
-
-    return validatedValue
-      ? toResolutionDetails(validatedValue, evaluationData)
-      : toResolutionDetails(defaultValue, evaluationData, StandardResolutionReasons.DEFAULT);
+    return this.evaluate(flagKey, 'boolean', context);
   }
 
   public async resolveStringEvaluation(
@@ -114,21 +96,7 @@ export class ConfigCatProvider implements Provider {
     defaultValue: string,
     context: EvaluationContext,
   ): Promise<ResolutionDetails<string>> {
-    if (!this._client) {
-      throw new GeneralError('Provider is not initialized');
-    }
-
-    const { value, ...evaluationData } = await this._client.getValueDetailsAsync<SettingValue>(
-      flagKey,
-      undefined,
-      context.targetingKey != null ? transformContext(context) : undefined,
-    );
-
-    const validatedValue = validateFlagType('string', value);
-
-    return validatedValue
-      ? toResolutionDetails(validatedValue, evaluationData)
-      : toResolutionDetails(defaultValue, evaluationData, StandardResolutionReasons.DEFAULT);
+    return this.evaluate(flagKey, 'string', context);
   }
 
   public async resolveNumberEvaluation(
@@ -136,21 +104,7 @@ export class ConfigCatProvider implements Provider {
     defaultValue: number,
     context: EvaluationContext,
   ): Promise<ResolutionDetails<number>> {
-    if (!this._client) {
-      throw new GeneralError('Provider is not initialized');
-    }
-
-    const { value, ...evaluationData } = await this._client.getValueDetailsAsync<SettingValue>(
-      flagKey,
-      undefined,
-      context.targetingKey != null ? transformContext(context) : undefined,
-    );
-
-    const validatedValue = validateFlagType('number', value);
-
-    return validatedValue
-      ? toResolutionDetails(validatedValue, evaluationData)
-      : toResolutionDetails(defaultValue, evaluationData, StandardResolutionReasons.DEFAULT);
+    return this.evaluate(flagKey, 'number', context);
   }
 
   public async resolveObjectEvaluation<U extends JsonValue>(
@@ -158,83 +112,49 @@ export class ConfigCatProvider implements Provider {
     defaultValue: U,
     context: EvaluationContext,
   ): Promise<ResolutionDetails<U>> {
+    const objectValue = await this.evaluate(flagKey, 'object', context);
+    return objectValue as ResolutionDetails<U>;
+  }
+
+  protected async evaluate<T extends PrimitiveTypeName>(
+    flagKey: string,
+    flagType: T,
+    context: EvaluationContext,
+  ): Promise<ResolutionDetails<PrimitiveType<T>>> {
     if (!this._client) {
-      throw new GeneralError('Provider is not initialized');
+      throw new ProviderNotReadyError('Provider is not initialized');
     }
 
     const { value, ...evaluationData } = await this._client.getValueDetailsAsync(
       flagKey,
       undefined,
-      context.targetingKey != null ? transformContext(context) : undefined,
+      transformContext(context),
     );
 
+    if (this._hasError && !evaluationData.errorMessage && !evaluationData.errorException) {
+      this._hasError = false;
+      this.events.emit(ProviderEvents.Ready);
+    }
+
     if (typeof value === 'undefined') {
-      return toResolutionDetails(defaultValue, evaluationData, StandardResolutionReasons.DEFAULT);
+      throw new FlagNotFoundError();
+    }
+
+    if (flagType !== 'object') {
+      return toResolutionDetails(flagType, value, evaluationData);
     }
 
     if (!isType('string', value)) {
-      throw new TypeMismatchError(`Requested object flag but the actual value is not a JSON string`);
+      throw new TypeMismatchError();
     }
 
+    let json: JsonValue;
     try {
-      const object = JSON.parse(value);
-
-      if (typeof object !== 'object') {
-        throw new TypeMismatchError(`Requested object flag but the actual value is ${typeof value}`);
-      }
-
-      return toResolutionDetails(object, evaluationData);
+      json = JSON.parse(value);
     } catch (e) {
-      if (e instanceof TypeMismatchError) {
-        throw e;
-      }
-
-      throw new ParseError(`Unable to parse '${value}' as JSON`);
+      throw new ParseError(`Unable to parse "${value}" as JSON`);
     }
+
+    return toResolutionDetails(flagType, json, evaluationData);
   }
-}
-
-function toResolutionDetails<U extends JsonValue>(
-  value: U,
-  data: Omit<IEvaluationDetails, 'value'>,
-  reason?: ResolutionReason,
-): ResolutionDetails<U> {
-  const matchedTargeting = 'matchedEvaluationRule' in data ? data.matchedEvaluationRule : data.matchedTargetingRule;
-  const matchedPercentage =
-    'matchedEvaluationPercentageRule' in data ? data.matchedEvaluationPercentageRule : data.matchedPercentageOption;
-
-  const matchedRule = Boolean(matchedTargeting || matchedPercentage);
-  const evaluatedReason = matchedRule ? StandardResolutionReasons.TARGETING_MATCH : StandardResolutionReasons.STATIC;
-
-  return {
-    value,
-    reason: reason ?? evaluatedReason,
-    errorMessage: data.errorMessage,
-    variant: data.variationId ?? undefined,
-  };
-}
-
-type PrimitiveTypeName = 'string' | 'boolean' | 'number' | 'object' | 'undefined';
-type PrimitiveType<T> = T extends 'string'
-  ? string
-  : T extends 'boolean'
-    ? boolean
-    : T extends 'number'
-      ? number
-      : T extends 'object'
-        ? object
-        : T extends 'undefined'
-          ? undefined
-          : unknown;
-
-function isType<T extends PrimitiveTypeName>(type: T, value: unknown): value is PrimitiveType<T> {
-  return typeof value === type;
-}
-
-function validateFlagType<T extends PrimitiveTypeName>(type: T, value: unknown): PrimitiveType<T> | undefined {
-  if (typeof value !== 'undefined' && !isType(type, value)) {
-    throw new TypeMismatchError(`Requested ${type} flag but the actual value is ${typeof value}`);
-  }
-
-  return value;
 }